working on the scraper
Just removed the puppeteer, and I have installed cheerio. I have also changed the core file for scraping StandardScraper; I saved the the old one on the same folder though. I am testing the new API for scraping on the "./__test" directory.  Scraping works but still some files would have to be changed in order to make sure the scripts are being read and executed accordingly by the application.