worked on the core class Scraper Script
Changed a lot of files and settings on the backend of the app. Finished setting up the new core ScraperScript class. Did a lot of testing with manually inserted data (no frontend.) It worked and it was blazingly fast; far faster than puppeteer. Will probably work on the controller for this class by tomorrow.