changed the configuration  of routes and controllers.
Separated handling of data (getting, posting, updating and deleting). Before the http response gets back to the client, we can inject other handlers, that can be anything, especially socket io. Also plan to add what was recommended before, which was to have a notification as to when new products are available for scraping. And to do that, I had to change the setup of the data handlers => any middleware / result handlers => httpRequestHandler (which returns the result to the client in JSON format).