work in progress
Changes in the app's dynamic creation of model, route, and script files. The app doesn't need to restart whenever we create or delete new route, model, and script files. The app just takes the dynamic route handler in memory and remove it after use... So if we happen to have deleted that route handler, the app wouldn't use it again even if it has never restarted.

Still working on the scraper controller, read, create and delete part of the scraper controller is done... just need to work on the update part, and the single scraper rewrite and the rewrite all functionality of the scraper.

Also did a lot of testing to make sure it works as intended.