work in progreses
I have just finished making the changes on the controller for the scraper - this is the file responsible for saving the scraper details on the database, writing the script module, the route module and the model module. This also means that the API route for the uploaded/saved scrapers are now working again.  Currently working on the controller that executes any of the saved scrapers / scripts.